+++ common/mptAlloc.h
@@ -79,19 +79,9 @@ struct GetRawBytesFunctor<const std::vector<T>>
 
 
 
-#if MPT_CXX_AT_LEAST(14)
 namespace mpt {
 using std::make_unique;
 } // namespace mpt
-#else
-namespace mpt {
-template<typename T, typename... Args>
-std::unique_ptr<T> make_unique(Args&&... args)
-{
-	return std::unique_ptr<T>(new T(std::forward<Args>(args)...));
-}
-} // namespace mpt
-#endif